+++ src/dawson.rs
@@ -4,13 +4,12 @@
 //! - `dawson`: Calculates Dawson's integral for real values.
 
 use crate::utils::sign;
-use std::f64::consts::PI;
+use core::f64::consts::PI;
 
 /// Dawson's integral for real values.
 /// Accurate to about 2e-7 for small inputs.
 pub fn dawson(x: f64) -> f64 {
     const NMAX: usize = 6;
-    let mut c = [0_f64; NMAX];
     let h = 0.4;
     let a1 = 2. / 3.;
     let a2 = 0.4;
@@ -26,10 +25,6 @@ pub fn dawson(x: f64) -> f64 {
     let xx: f64;
     let ans: f64;
 
-    for i in 0..NMAX {
-        c[i] = f64::exp(-((2. * i as f64 + 1.) * h).powf(2.));
-    }
-
     if x.abs() < 0.2 {
         x2 = x * x;
         ans = x * (1.0 - a1 * x2 * (1.0 - a2 * x2 * (1.0 - a3 * x2)));
@@ -44,14 +39,15 @@ pub fn dawson(x: f64) -> f64 {
         sum = 0.0;
 
         for i in 0..NMAX {
+            let cval = f64::exp(-((2. * i as f64 + 1.) * h).powi(2));
             d1 += 2.;
             d2 -= 2.;
             e1 *= e2;
 
-            sum += c[i] * (e1 / d1 + 1.0 / (d2 * e1));
+            sum += cval * (e1 / d1 + 1.0 / (d2 * e1));
         }
 
         ans = (1. / PI.sqrt()) * sign(f64::exp(-xp * xp), x) * sum;
     }
-    return ans;
+    ans
 }